Fix uint8 pipeline to achieve 4x GPU bandwidth savings (#434)
## Summary

- **Fixes incomplete uint8 pipeline** from PR #431 that was still
converting images to float32 on CPU
- **Achieves the promised 4x GPU bandwidth savings** by keeping images
in uint8 until GPU normalization
- **Removes unnecessary CPU normalization** from all 6 dataset classes

## Background

PR #431 added Skia-based augmentation and claimed:
> "Reduces GPU transfer bandwidth by **4x** (images stay uint8 until GPU
normalization)"

However, the datasets still called `apply_normalization()` which
converted uint8→float32 **on CPU**, negating the bandwidth savings. The
`normalize_on_gpu()` function was added but ended up being a no-op since
images were already float32.

## Changes

### `sleap_nn/data/custom_datasets.py`
- Remove `apply_normalization()` calls from all 6 dataset classes:
  - `BottomUpDataset`
  - `BottomUpMultiClassDataset`
  - `CenteredInstanceDataset`
  - `TopDownCenteredInstanceMultiClassDataset`
  - `SingleInstanceDataset`
  - `CentroidDataset`

### `sleap_nn/data/resizing.py`
- Remove `.to(torch.float32)` from `apply_pad_to_stride()` and
`apply_sizematcher()`
- These were converting uint8 images to float32 during padding
operations

### `tests/data/test_custom_datasets.py`
- Add 4 new tests verifying uint8 output from datasets:
  - `test_uint8_pipeline_bottomup`
  - `test_uint8_pipeline_singleinstance`
  - `test_uint8_pipeline_centroid`
  - `test_uint8_pipeline_centered_instance`

## Pipeline Flow

**Before (PR #431):**
```
process_lf() → uint8 → apply_normalization() → float32 → [GPU transfer 4 bytes/pixel]
                                                              ↓
                                     normalize_on_gpu() (NO-OP, already float32)
```

**After (this PR):**
```
process_lf() → uint8 → [GPU transfer 1 byte/pixel] → normalize_on_gpu() → float32
```

## Test Results

- All 56 data tests pass
- All 157 training tests pass
- 4 new uint8 pipeline tests pass

## Test plan

- [x] Run `pytest tests/data/test_custom_datasets.py` - all 11 tests
pass
- [x] Run `pytest tests/data/` - all 56 tests pass
- [x] Run `pytest tests/training/` - all 157 tests pass
- [ ] Verify bandwidth savings with profiling (optional)


---------

+++ sleap_nn/data/resizing.py
@@ -63,7 +63,7 @@ def apply_pad_to_stride(image: torch.Tensor, max_stride: int) -> torch.Tensor:
                 image,
                 (0, pad_width, 0, pad_height),
                 mode="constant",
-            ).to(torch.float32)
+            )
     return image
 
 
@@ -136,7 +136,7 @@ def apply_sizematcher(
             image,
             (0, pad_width, 0, pad_height),
             mode="constant",
-        ).to(torch.float32)
+        )
 
         return image, eff_scale_ratio
     else:
